made creating list conditional based on if there are values in the array
We removed everything except for the list we were working on so it would be easier to focus on.  We created a method which creates an unordered list if the array has values and displays a paragraph if there are not values in the array. We also used && to display a message is the array is empty. We deled the values from the array to test these.